requirements.txt
tinytag
ffmpeg-python
Pillow